Create timedSpawn
In this code, we create a Player class that inherits from pygame.sprite.Sprite. The Player class has an update() method that we don't need to use in this example, so we leave it empty. We create a Player object and add it to a pygame.sprite.Group object called player_group. In the game loop, we handle events as usual. When a MOUSEBUTTONDOWN event occurs, we check if the mouse click collided with the player sprite using the collidepoint() method of the player's rect attribute. If there was a collision, we remove the player sprite from the player_group using the kill() method. We then update and draw the player sprite as usual, and limit the frame rate to 60 FPS using clock.tick(60). With this code, when you click on the player sprite, it will disappear from the screen.